+++ internal/services/userevents/mocks_test.go
@@ -296,7 +296,7 @@ func (m *MockUserUsedSpaceEventHandler) EXPECT() *MockUserUsedSpaceEventHandlerM
 }
 
 // HandleUsedSpaceEvent mocks base method.
-func (m *MockUserUsedSpaceEventHandler) HandleUsedSpaceEvent(arg0 context.Context, arg1 int) error {
+func (m *MockUserUsedSpaceEventHandler) HandleUsedSpaceEvent(arg0 context.Context, arg1 int64) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "HandleUsedSpaceEvent", arg0, arg1)
 	ret0, _ := ret[0].(error)
